requirements.txt
streamlit
pandas
plotly
sqlalchemy
pymysql
mysql-connector-python
python-dotenv
tweepy
vaderSentiment
mysql-connector-python
SQLAlchemy
python-dotenv
nltk
textblob
scikit-learn
SQLAlchemy
nltk
textblob
scikit-learn
pip install textblob nltk
python -m textblob.download_corpora
pip install SpeechRecognition
pip install pyaudio
# Core Framework
streamlit==1.28.0
flask==2.3.3

# Database
mysql-connector-python==8.1.0
SQLAlchemy==2.0.21
python-dotenv==1.0.0

# Authentication
bcrypt==4.0.1

# Web Scraping
selenium==4.13.0
beautifulsoup4==4.12.2
requests==2.31.0
scrapy==2.11.0

# Social Media APIs
tweepy==4.14.0
instaloader==4.9.6

# NLP & ML
transformers==4.33.2
torch==2.0.1
scikit-learn==1.3.1
nltk==3.8.1
spacy==3.6.1

# Sentiment Analysis Models
sentencepiece==0.1.99

# Translation & Language Detection
googletrans==4.0.0rc1
langdetect==1.0.9

# Voice Recognition
SpeechRecognition==3.10.0
pyaudio==0.2.13

# Image Analysis
Pillow==10.0.1
opencv-python==4.8.1.78

# CLIP for Image Understanding
ftfy==6.1.1
regex==2023.8.8

# Data Visualization
plotly==5.17.0
pandas==2.1.1
numpy==1.25.2
matplotlib==3.8.0
seaborn==0.12.2

# Geographic Analysis
folium==0.14.0
geopy==2.4.0
pip install branca

# Text Processing
keybert==0.8.3
rake-nltk==1.0.6

# Utilities
tqdm==4.66.1
python-dateutil==2.8.2

# Excel/CSV Processing
openpyxl==3.1.2
xlrd==2.0.1

# API Development
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.4.2

# Testing
pytest==7.4.2
pytest-cov==4.1.0

# Linting
flake8==6.1.0
black==23.9.1

# Jupyter (for development)
jupyter==1.0.0
ipython==8.15.0
pip install torch==2.5.1

# Core dependencies
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Selenium for web scraping
selenium>=4.15.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# NLP libraries
vaderSentiment>=3.3.2
textblob>=0.17.1

# Optional: Advanced NLP (large download)
# transformers>=4.35.0
# torch>=2.0.0

# Social media APIs
tweepy>=4.14.0
instaloader>=4.10.0

# Environment management
python-dotenv>=1.0.0

# Web framework (for dashboard)
streamlit>=1.28.0
plotly>=5.17.0

# Database (for caching)
sqlalchemy>=2.0.0

# Logging and utilities
colorama>=0.4.6
tqdm>=4.66.0
python cli_analyzer.py --url "https://www.amazon.in/dp/B08N5WRWNW" --max 50 --output results.csv
python -m pip install sqlalchemy --upgrade --force-reinstall


